suppression des nuls dans les requetes not in

+++ scripts/cleanGroup.pl
@@ -107,7 +107,7 @@ print "$cpt group user supprimés\n";
 # on delete les tuples de oc_asso_uai_user_group inutiles
 
 my @sqlQueries = ("delete from oc_asso_uai_user_group where exists (select * from oc_preferences where  user_group =userid and appid = 'core' and configkey = 'enabled' and configvalue = 'false')",
-			"delete from oc_asso_uai_user_group a where user_group not in (select gid from oc_groups) and user_group not in (select uid from oc_users)");
+			"delete from oc_asso_uai_user_group a where user_group not in (select gid from oc_groups where gid is not null) and user_group not in (select uid from oc_users where uid is not null)");
 
 if ($all) {
 	my $nb = 0;
